requirements.txt
streamlit
streamlit-audiorecorder
streamlit-audiorec
openai-whisper